Replacing RefCell with UnsafeCell.
Using RefCell to implement the SubCell was a bad idea. Indeed, if
several readers attempt to borrow() the SubCell concurrently, they
will mutate concurrently the refcount of the RefCell, which is
entirely thread-unsafe. Panic ensues.

Since we don't actually need the r/w provided by RefCell (it is
already guaranteed by the MainLock), this patch entirely discards the
RefCell and replaces it by an UnsafeCell.